+++ plugins/postcss-nesting/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### Unreleased
 
-- Fix semicolons being removed (https://github.com/csstools/postcss-plugins/issues/497).
+- Fix selector order with any pseudo element.
+- Fix semicolons being removed [#497](https://github.com/csstools/postcss-plugins/issues/497).
 
 ### 10.1.8 (June 4, 2022)